products added

+++ app/src/main/java/com/example/mcdonaldsclone/features/menu/model/Product.kt
@@ -7,7 +7,7 @@ data class Product(
     val name: String,
     @DrawableRes val imageResId: Int,
     val price: Double,
-    val categoryId: Int,
+    val categoryId: Long,
     val subCategoryId: Long,
     val isSetAvailable: Boolean,
     val setPrice: Double = 0.00,
